go for it ... changed driving controller to less bandwidth-consuming transmission

+++ thor_mang_driving_controller/include/thor_mang_driving_controller/thor_mang_driving_controller.h
@@ -88,9 +88,9 @@ private:
   bool controller_enabled_;
 
   // accumulate absolute steering angle
-  double current_absolute_steering_angle_;
-  double current_head_tilt_;
-  double current_head_pan_;
+  float current_absolute_steering_angle_;
+  float current_head_tilt_;
+  float current_head_pan_;
   unsigned int driving_counter_;
 
   double head_sensitivity_;
